requirements.txt
streamlit
groq
requests
Pillow
pytesseract
